refactor: 인증관련 controller메서드는 AuthController로 이동
